+++ tests/testthat/test-ggbetweenstats.R
@@ -131,17 +131,15 @@ testthat::test_that(
         displaystyle(atop(
           displaystyle("From ggplot2 package"),
           expr = paste(
-            "log"["e"],
-            "(BF"["01"],
-            ") = ",
-            "1.54274"
+            "log"["e"], "(BF"["01"],
+            ") = ", "1.54274"
           )
         )),
         expr = paste(
-          "Pairwise comparisons: ",
+          "Pairwise test: ",
           bold("Games-Howell test"),
-          "; Adjustment (p-value): ",
-          bold("Holm")
+          "; Comparisons shown: ",
+          bold("only significant")
         )
       ))
     )
@@ -166,6 +164,7 @@ testthat::test_that(
         type = "np",
         mean.ci = TRUE,
         k = 2L,
+        pairwise.comparisons = FALSE,
         conf.level = 0.90,
         outlier.tagging = TRUE,
         outlier.label = "name",